requirements.txt
Appium-Python-Client==0.24
py==1.4.34
pytest==3.2.3
selenium==3.7.0
